Fix bug in download_files function
I notice that often the program doesn't download all the file it claims to have found, so I investigated a bit and I found that Pymeta was setting as user-agent a random character due to the fact that there were the choice function applied to what random_agent() returns, that is a string and not a list.
I also think that there is a missing sleep in case of more than the max_results results are found. Maybe it can help to avoid being captcha'ed.